+++ src/app/admin/containers/monitors/monitors.component.ts
@@ -6,6 +6,7 @@ import { ToastrService } from 'ngx-toastr';
 import { MonitorService } from '../../monitor.service';
 import { Monitor } from '../../models/monitor';
 import { MonitorsDialogComponent } from '../../components/monitors-dialog/monitors-dialog.component';
+import { ConfirmDialogComponent } from '../../../shared/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-monitors',
@@ -59,12 +60,23 @@ export class MonitorsComponent implements OnInit {
   }
 
   deleteMonitor(monitor: Monitor) {
-    this.monitorsService.deleteMonitor(monitor.id_monitor).subscribe(data => {
-      if (data && data['affectedRows'] > 0) {
-        this.toastr.success('Operación realizada con éxito', 'El monitor ha sido eliminado');
-        this.getMonitors();
-      } else {
-        this.toastr.error('Algo ha fallado en la operación', 'Inténtelo más tarde');
+    const initialState = {
+      title: 'Eliminar monitor',
+      content: '¿Está seguro que desea borrar a este monitor? Se eliminará permanentemente del sistema',
+      class: 'modal-lg'
+    };
+    this.bsModalRef = this.modalService.show(ConfirmDialogComponent, {initialState});
+    this.bsModalRef.content.closeBtnName = 'Cerrar';
+    this.modalService.onHide.subscribe(useless => {
+      if (this.bsModalRef.content.borrar) {
+        this.monitorsService.deleteMonitor(monitor.id_monitor).subscribe(data => {
+          if (data && data['affectedRows'] > 0) {
+            this.toastr.success('Operación realizada con éxito', 'El monitor ha sido eliminado');
+            this.getMonitors();
+          } else {
+            this.toastr.error('Algo ha fallado en la operación', 'Inténtelo más tarde');
+          }
+        }); 
       }
     });
   }
